Extract config inject logic
